pyproject.toml
[tool.poetry]
name = "camera-presence"
version = "0.1.0"
description = ""
authors = ["Foo"]
readme = "README.md"
packages = [{include = "camera_presence"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
opencv-python = "^4.6.0.66"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = 'camera_presence.__main__:main'
